instalock/requirements.txt
screeninfo~=0.8.1
Pillow~=9.5.0
setuptools~=65.5.1
PyAutoGUI~=0.9.53
opencv-python~=4.7.0.72
numpy~=1.24.2